+++ src/mainzone.js
@@ -7,7 +7,7 @@ import { PictureModesConversionToHomeKit, PictureModesDenonNumber, PictureModesD
 let Accessory, Characteristic, Service, Categories, Encode, AccessoryUUID;
 
 class MainZone extends EventEmitter {
-    constructor(api, device, zone, name, host, port, generation, devInfoFile, inputsFile, inputsNamesFile, inputsTargetVisibilityFile, refreshInterval) {
+    constructor(api, device, name, host, port, generation, zone, devInfoFile, inputsFile, inputsNamesFile, inputsTargetVisibilityFile) {
         super();
 
         Accessory = api.platformAccessory;
@@ -34,14 +34,14 @@ class MainZone extends EventEmitter {
         this.sensorMute = device.sensorMute || false;
         this.sensorInput = device.sensorInput || false;
         this.sensorInputs = device.sensorInputs || [];
-        this.masterPower = device.masterPower || false;
+        this.powerControlZone = device.powerControlZone || 0;
         this.volumeControlNamePrefix = device.volumeControlNamePrefix || false;
         this.volumeControlName = device.volumeControlName || 'Volume';
         this.volumeControlType = device.volumeControlType || 0;
         this.volumeControlZone = device.volumeControlZone || 0;
         this.volumeMax = device.volumeMax || 100;
         this.infoButtonCommand = device.infoButtonCommand || 'MNINF';
-        this.refreshInterval = refreshInterval;
+        this.refreshInterval = device.refreshInterval * 1000 || 5000;
         this.enableDebugMode = device.enableDebugMode || false;
         this.disableLogInfo = device.disableLogInfo || false;
         this.devInfoFile = devInfoFile;
@@ -147,8 +147,8 @@ class MainZone extends EventEmitter {
             let set = false
             switch (key) {
                 case 'Power':
-                    const powerState = this.masterPower ? (value ? 'PWON' : 'PWSTANDBY') : (value ? 'ZMON' : 'ZMOFF');
-                    set = await this.denon.send(powerState)
+                    const powerState = value ? 'ON' : 'OFF';
+                    set = await this.stateControl('Power', powerState);
                     break;
                 case 'Input':
                     const input = `SI${value}`;
@@ -160,11 +160,11 @@ class MainZone extends EventEmitter {
                     break;
                 case 'Volume':
                     const volume = (value < 0 || value > 100) ? this.volume : (value < 10 ? `0${value}` : value);
-                    set = await this.volumeControl('Volume', volume);
+                    set = await this.stateControl('Volume', volume);
                     break;
                 case 'Mute':
                     const mute = value ? 'ON' : 'OFF';
-                    set = await this.volumeControl('Mute', mute);
+                    set = await this.stateControl('Mute', mute);
                     break;
                 case 'RcControl':
                     set = await this.denon.send(value);
@@ -322,142 +322,52 @@ class MainZone extends EventEmitter {
         return scaledValue;
     }
 
-    async volumeControl(type, value) {
+    async stateControl(type, value) {
         try {
-            switch (type) {
-                case 'VolumeSelector': //VolumeSelector
-                    switch (this.volumeControlZone) {
-                        case 0:
-                            value = `MV${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 1:
-                            value = `Z2${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 2:
-                            value = `Z3${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 3:
-                            value = `Z2${value}`;
-                            await this.denon.send(value);
-                            value = `Z3${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 4:
-                            value = `MV${value}`;
-                            await this.denon.send(value);
-                            value = `Z2${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 5:
-                            value = `MV${value}`;
-                            await this.denon.send(value);
-                            value = `Z3${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 6:
-                            value = `MV${value}`;
-                            await this.denon.send(value);
-                            value = `Z2${value}`;
-                            await this.denon.send(value);
-                            value = `Z3${value}`;
-                            await this.denon.send(value);
-                            break;
-                    }
-                    break;
-                case 'Volume': //Volume
-                    switch (this.volumeControlZone) {
-                        case 0:
-                            value = `MV${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 1:
-                            value = `Z2${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 2:
-                            value = `Z3${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 3:
-                            value = `Z2${value}`;
-                            await this.denon.send(value);
-                            value = `Z3${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 4:
-                            value = `MV${value}`;
-                            await this.denon.send(value);
-                            value = `Z2${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 5:
-                            value = `MV${value}`;
-                            await this.denon.send(value);
-                            value = `Z3${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 6:
-                            value = `MV${value}`;
-                            await this.denon.send(value);
-                            value = `Z2${value}`;
-                            await this.denon.send(value);
-                            value = `Z3${value}`;
-                            await this.denon.send(value);
-                            break;
-                    }
-                    break;
-                case 'Mute': //Mute
-                    switch (this.volumeControlZone) {
-                        case 0:
-                            value = `MU${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 1:
-                            value = `Z2MU${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 2:
-                            value = `Z3MU${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 3:
-                            value = `Z2MU${value}`;
-                            await this.denon.send(value);
-                            value = `Z3MU${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 4:
-                            value = `MU${value}`;
-                            await this.denon.send(value);
-                            value = `Z2MU${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 5:
-                            value = `MU${value}`;
-                            await this.denon.send(value);
-                            value = `Z3MU${value}`;
-                            await this.denon.send(value);
-                            break;
-                        case 6:
-                            value = `MU${value}`;
-                            await this.denon.send(value);
-                            value = `Z2MU${value}`;
-                            await this.denon.send(value);
-                            value = `Z3MU${value}`;
-                            await this.denon.send(value);
-                            break;
-                    }
-                    break;
-                default:
-                    this.emit('warn', `unknown volume control type: ${type}`);
-                    break;
+            // Normalize value for Power type
+            if (type === 'Power' && value === 'OFF' && this.volumeControlZone === 7) {
+                value = 'STANDBY';
             }
+
+            // Define main zone
+            const mainZone = type === 'Power' ? 'ZM' : 'MV';
+            const zoneMap = {
+                0: [mainZone],
+                1: ['Z2'],
+                2: ['Z3'],
+                3: ['Z2', 'Z3'],
+                4: [mainZone, 'Z2'],
+                5: [mainZone, 'Z3'],
+                6: [mainZone, 'Z2', 'Z3'],
+                7: ['PW']
+            };
+
+            // Reuse volume control zones for better readability
+            const typeMap = {
+                'Power': zoneMap[this.powerControlZone],
+                'VolumeSelector': zoneMap[this.volumeControlZone],
+                'Volume': zoneMap[this.volumeControlZone],
+                'Mute': zoneMap[this.volumeControlZone].map(zone => zone ? `${zone}MU` : 'MU')
+            };
+
+            // Get the commands for the specified type
+            const commands = typeMap[type];
+            if (commands) {
+                const commandsCount = commands.length;
+                for (let i = 0; i < commandsCount; i++) {
+                    const cmd = commands[i];
+                    await this.denon.send(`${cmd}${value}`);
+                    const pauseTime = type === 'Power' && value === 'ON' && commandsCount > 1 && i === 0 ? 4000 : 75;
+                    const pause = i < commandsCount - 1 ? await new Promise(resolve => setTimeout(resolve, pauseTime)) : false;
+                }
+            } else {
+                this.emit('warn', `Unknown control type: ${type}`);
+            }
+
+            return true;
         } catch (error) {
-            this.emit('warn', `volume control error: ${error}`);
-        };
+            this.emit('warn', `State control error for type ${type} with value ${value}: ${error}`);
+        }
     }
 
     //prepare accessory
@@ -498,8 +408,8 @@ class MainZone extends EventEmitter {
                     }
 
                     try {
-                        const powerState = this.masterPower ? (state ? 'PWON' : 'PWSTANDBY') : (state ? 'ZMON' : 'ZMOFF');
-                        await this.denon.send(powerState);
+                        const powerState = state ? 'ON' : 'OFF';
+                        await this.stateControl('Power', powerState);
                         const info = this.disableLogInfo ? false : this.emit('info', `set Power: ${powerState}`);
                     } catch (error) {
                         this.emit('warn', `set Power error: ${error}`);
@@ -690,11 +600,11 @@ class MainZone extends EventEmitter {
                         switch (command) {
                             case Characteristic.VolumeSelector.INCREMENT:
                                 command = 'UP';
-                                await this.volumeControl('VolumeSelector', command);
+                                await this.stateControl('VolumeSelector', command);
                                 break;
                             case Characteristic.VolumeSelector.DECREMENT:
                                 command = 'DOWN';
-                                await this.volumeControl('VolumeSelector', command);
+                                await this.stateControl('VolumeSelector', command);
                                 break;
                         }
                         const info = this.disableLogInfo ? false : this.emit('info', `set Volume Selector: ${command}`);
@@ -712,7 +622,7 @@ class MainZone extends EventEmitter {
                     try {
                         let scaledValue = await this.scaleValue(value, 0, 100, 0, this.volumeMax >= 2 ? this.volumeMax - 2 : this.volumeMax);
                         scaledValue = scaledValue < 10 ? `0${scaledValue}` : scaledValue;
-                        await this.volumeControl('Volume', scaledValue);
+                        await this.stateControl('Volume', scaledValue);
                         const info = this.disableLogInfo ? false : this.emit('info', `set Volume: ${value}%`);
                     } catch (error) {
                         this.emit('warn', `set Volume error: ${error}`);
@@ -727,7 +637,7 @@ class MainZone extends EventEmitter {
                 .onSet(async (state) => {
                     try {
                         state = state ? 'ON' : 'OFF';
-                        await this.volumeControl('Mute', state);
+                        await this.stateControl('Mute', state);
                         const info = this.disableLogInfo ? false : this.emit('info', `set Mute: ${state}`);
                     } catch (error) {
                         this.emit('warn', `set Mute error: ${error}`);
@@ -756,8 +666,8 @@ class MainZone extends EventEmitter {
                 const name = input.name ?? `Input ${inputIdentifier}`;
 
                 //get input name
-                const savedInputsNames = this.savedInputsNames[inputReference] ?? false;
-                input.name = savedInputsNames ? savedInputsNames.substring(0, 64) : name.substring(0, 64);
+                const savedInputsName = this.savedInputsNames[inputReference] ?? false;
+                input.name = savedInputsName ? savedInputsName.substring(0, 64) : name.substring(0, 64);
 
                 //get type
                 const inputSourceType = 0;
@@ -966,7 +876,7 @@ class MainZone extends EventEmitter {
                     this.allServices.push(sensorInputService);
                     accessory.addService(sensorInputService);
                 }
-            }
+            };
 
             //prepare buttons services
             const possibleButtonsCount = 99 - this.allServices.length;
@@ -1008,6 +918,7 @@ class MainZone extends EventEmitter {
                                 const reference = command;
 
                                 const set = state ? await this.denon.send(reference) : false;
+                                const set1 = state && directSound ? await new Promise(resolve => setTimeout(resolve, 75)) : false;
                                 const set2 = state && directSound ? await this.denon.send(directSoundModeSurround) : false;
                                 const info = this.disableLogInfo || !state ? false : this.emit('info', `set Button Name: ${buttonName}, Reference: ${reference}`);
                             } catch (error) {